refactor: Remove unused file input and handle file upload in FilesMenu component
Remove the unused file input and handle file upload logic from the ComposerToolbar component and refactor it into the new FilesMenu component. This change improves code organization and separates concerns by moving the file upload functionality to a dedicated component. It also removes unnecessary code and simplifies the ComposerToolbar component.
